controller: track: Fix delete route to understand user id info
Fix the track route to be post, take a track ID as input, and also to
consider both if the user is logged in and if they have permission to
delete the track before actually deleting it, and return a 401 error
otherwise.

+++ app/controller/api/v1/track.controller.php
@@ -80,9 +80,21 @@ class TrackController extends Controller
 
     public function delete()
     {
-        $track = new TrackModel();
+        if (!$this->logged_in) {
+            $response = json_response_fail(401, [], "User Not Authorized");
+            $this->data('response', $response);
+            return;
+        }
 
-        $id = $this->params['id'];
+        $id = $this->params['json']['id'];
+        $user_id = $this->user_data->id;
+
+        $track = TrackModel::find($id);
+        if ($track->user_id != $user_id) {
+            $response = json_response_fail(401, [], "User Not Authorized");
+            $this->data('response', $response);
+            return;
+        }
 
         $track->delete("id = \"" . $id. "\"");
